refactor classification by id
