+++ klass-api/src/test/java/no/ssb/klass/api/migration/dataintegrity/classification/KlassApiClassificationByIdXmlTest.java
@@ -1,0 +1,121 @@
+package no.ssb.klass.api.migration.dataintegrity.classification;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
+
+import static no.ssb.klass.api.migration.MigrationTestConstants.APPLICATION_XML;
+import static no.ssb.klass.api.migration.MigrationTestConstants.CLASSIFICATIONS_PATH;
+import static no.ssb.klass.api.migration.MigrationTestUtils.*;
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+
+public class KlassApiClassificationByIdXmlTest extends AbstractKlassApiClassificationTest{
+
+
+    @Test
+    void getOneClassificationXml() throws Exception {
+
+        int classificationId = sourceResponseIdentifiers.get(11);
+        String path = getClassificationByIdPath(classificationId);
+
+        sourceResponse = klassApiMigrationClient.getFromSourceApi(path, null, APPLICATION_XML);
+        targetResponse = klassApiMigrationClient.getFromTargetApi(path, null, APPLICATION_XML);
+
+        assertApiResponseIsNotNull(sourceResponse);
+
+        assertStatusCodesEqual(sourceResponse.getStatusCode(), targetResponse.getStatusCode(), path);
+
+        if(sourceResponse.getStatusCode() != 200) {
+            assertThat(compareError(classificationId, sourceResponse, targetResponse)).isTrue();
+        }
+        else {
+            validateObjectXml(path, sourceResponse, targetResponse);
+            validateLinksXml(path,sourceResponse, targetResponse);
+        }
+    }
+
+    @ParameterizedTest
+    @MethodSource("rangeProviderClassificationIds")
+    void getClassificationXml(Integer classificationId) throws Exception {
+
+        String path = getClassificationByIdPath(classificationId);
+
+        sourceResponse = klassApiMigrationClient.getFromSourceApi(path, null, APPLICATION_XML);
+        targetResponse = klassApiMigrationClient.getFromTargetApi(path, null, APPLICATION_XML);
+
+        assertApiResponseIsNotNull(sourceResponse);
+
+        assertStatusCodesEqual(sourceResponse.getStatusCode(), targetResponse.getStatusCode(), path);
+
+        if(sourceResponse.getStatusCode() != 200) {
+            assertThat(compareError(classificationId, sourceResponse, targetResponse)).isTrue();
+        }
+        else {
+            validateObjectXml(path, sourceResponse, targetResponse);
+            validateLinksXml(path,sourceResponse, targetResponse);
+        }
+    }
+
+    @ParameterizedTest
+    @MethodSource("rangeProviderClassificationIds")
+    void getClassificationEnglishXml(Integer classificationId) throws Exception {
+
+        String path = getClassificationByIdPath(classificationId);
+
+        sourceResponse = klassApiMigrationClient.getFromSourceApi(path, paramsLanguageEn, APPLICATION_XML);
+        targetResponse = klassApiMigrationClient.getFromTargetApi(path, paramsLanguageEn, APPLICATION_XML);
+
+        assertApiResponseIsNotNull(sourceResponse);
+
+        assertStatusCodesEqual(sourceResponse.getStatusCode(), targetResponse.getStatusCode(), path);
+
+        if(sourceResponse.getStatusCode() != 200) {
+            assertThat(compareError(classificationId, sourceResponse, targetResponse)).isTrue();
+        } else {
+            validateObjectXml(path, sourceResponse, targetResponse);
+            validateLinksXml(path,sourceResponse, targetResponse);
+        }
+    }
+
+    @ParameterizedTest
+    @MethodSource("rangeProviderClassificationIds")
+    void getClassificationNewNorwegianXml(Integer classificationId) throws Exception {
+
+        String path = CLASSIFICATIONS_PATH + "/" + classificationId;
+
+        sourceResponse = klassApiMigrationClient.getFromSourceApi(path, paramsLanguageNn, APPLICATION_XML);
+        targetResponse = klassApiMigrationClient.getFromTargetApi(path, paramsLanguageNn, APPLICATION_XML);
+
+        assertApiResponseIsNotNull(sourceResponse);
+
+        assertStatusCodesEqual(sourceResponse.getStatusCode(), targetResponse.getStatusCode(), path);
+
+        if(sourceResponse.getStatusCode() != 200) {
+            assertThat(compareError(classificationId, sourceResponse, targetResponse)).isTrue();
+        }
+        else {
+            validateObjectXml(path, sourceResponse, targetResponse);
+            validateLinksXml(path,sourceResponse, targetResponse);
+        }
+    }
+
+    @ParameterizedTest
+    @MethodSource("rangeProviderClassificationIds")
+    void getClassificationIncludeFutureXml(Integer classificationId) throws Exception {
+
+        String path = getClassificationByIdPath(classificationId);
+        sourceResponse = klassApiMigrationClient.getFromSourceApi(path, paramsIncludeFuture, APPLICATION_XML);
+        targetResponse = klassApiMigrationClient.getFromTargetApi(path, paramsIncludeFuture, APPLICATION_XML);
+
+        assertApiResponseIsNotNull(sourceResponse);
+
+        assertStatusCodesEqual(sourceResponse.getStatusCode(), targetResponse.getStatusCode(), path);
+
+        if(sourceResponse.getStatusCode() != 200) {
+            assertThat(compareError(classificationId, sourceResponse, targetResponse)).isTrue();
+        }else {
+            validateObjectXml(path, sourceResponse, targetResponse);
+            validateLinksXml(path,sourceResponse, targetResponse);
+        }
+    }
+}
